Added 'deleted_at' and 'parent' field to objects.
projects, activities, and times now have the 'deleted_at' and 'parent'
field for storing if they are an updated object ('parent' refers to the
original object) and if the object is deleted_at (which refers to if an
object has been archived or not).

There were also some miscellaneous formatting changes.